requirements.txt
strands-agents>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pytest>=8.0.0
